Sequential execution instead of parallel that leads to rate limits
